preliminary cleanup
